refactor: Cafe DTO, Service 패키지 변경
