+++ src/main/java/com/example/demo/dto/cafe/CafeBasicInfoResponse.java
@@ -1,4 +1,4 @@
-package com.example.demo.dto;
+package com.example.demo.dto.cafe;
 
 import java.util.List;
