+++ src/main/java/com/example/demo/dto/cafe/CafeResponse.java
@@ -1,6 +1,10 @@
-package com.example.demo.dto;
+package com.example.demo.dto.cafe;
 
 import java.util.List;
+
+import com.example.demo.dto.CanMakeStudyOnceResponse;
+import com.example.demo.dto.ReviewResponse;
+import com.example.demo.dto.StudyOnceForCafeResponse;
 
 import lombok.Builder;
 import lombok.Getter;
