+++ src/main/java/com/example/demo/dto/cafe/CafeSearchRequest.java
@@ -1,4 +1,6 @@
-package com.example.demo.dto;
+package com.example.demo.dto.cafe;
+
+import com.example.demo.dto.PagedRequest;
 
 import lombok.Getter;
 import lombok.Setter;
