+++ src/main/java/com/example/demo/mapper/BusinessHourMapper.java
@@ -4,7 +4,7 @@ import java.util.List;
 import java.util.stream.Collectors;
 
 import com.example.demo.domain.cafe.BusinessHour;
-import com.example.demo.dto.BusinessHourResponse;
+import com.example.demo.dto.cafe.BusinessHourResponse;
 
 public class BusinessHourMapper {
