+++ src/main/java/com/example/demo/service/cafe/CafeQueryService.java
@@ -1,9 +1,9 @@
-package com.example.demo.service;
+package com.example.demo.service.cafe;
 
-import com.example.demo.dto.CafeResponse;
-import com.example.demo.dto.CafeSearchRequest;
-import com.example.demo.dto.CafeSearchResponse;
 import com.example.demo.dto.PagedResponse;
+import com.example.demo.dto.cafe.CafeResponse;
+import com.example.demo.dto.cafe.CafeSearchRequest;
+import com.example.demo.dto.cafe.CafeSearchResponse;
 
 public interface CafeQueryService {
