+++ src/main/java/com/example/demo/service/cafe/CafeService.java
@@ -1,0 +1,4 @@
+package com.example.demo.service.cafe;
+
+public interface CafeService {
+}
