+++ src/main/java/com/example/demo/service/cafe/CafeServiceImpl.java
@@ -1,4 +1,4 @@
-package com.example.demo.service;
+package com.example.demo.service.cafe;
 
 import org.springframework.stereotype.Service;
